+++ lib/constants.ts
@@ -1,2 +1,4 @@
 export const WEBSITE_URL = 'https://yaps.gg'
 export const SUBSCRIPTION_URL = "https://www.patreon.com/abdibrokhim/membership"
+export const REPO_URL_FORK = "https://github.com/abdibrokhim/yaps.gg-shadcnui/fork"
+export const REPO_URL_STAR = "https://github.com/abdibrokhim/yaps.gg-shadcnui"